api moved under app, added probablistic output for api, added react app for frontend

+++ app/api/app.py
@@ -1,4 +1,5 @@
 from fastapi import FastAPI, File, UploadFile
+from fastapi.middleware.cors import CORSMiddleware
 from AudioService import AudioService
 from ModelService import ModelService
 import librosa
@@ -17,8 +18,16 @@ app = FastAPI(
 audio_service = AudioService()
 model_service = ModelService()
 
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"]
+)
 
-@app.post("/upload")
+
+@app.post("/final_prediction")
 def test_upload(file: UploadFile = File()):
     if not file:
         return {"message": "ERROR! No upload file sent"}
@@ -40,6 +49,35 @@ def test_upload(file: UploadFile = File()):
 
         spectrograms = audio_service.get_spectrograms_from_stream(stream, sample_rate)
         prediction_dict = model_service.classify_audio(spectrograms)
+
+        tf.close()
+        os.unlink(tf.name)
+        return prediction_dict
+    except Exception as e:
+        return {"message": f"ERROR! File could not be read. \nException: {e}"}
+
+@app.post("/window_predictions")
+def test_upload(file: UploadFile = File()):
+    if not file:
+        return {"message": "ERROR! No upload file sent"}
+    try:
+        tf = tempfile.NamedTemporaryFile(delete=False)
+        tfName = tf.name
+        tf.seek(0)
+        tf.write(file.file.read())
+        tf.flush()
+
+        sample_rate = librosa.get_samplerate(tfName)
+
+        stream = librosa.stream(tfName,
+                                block_length=1,
+                                frame_length=3 * int(sample_rate),
+                                hop_length=int(sample_rate),
+                                fill_value=0,
+                                mono=True)
+
+        spectrograms = audio_service.get_spectrograms_from_stream(stream, sample_rate)
+        prediction_dict = model_service.classify_audio_window_probabilities(spectrograms)
 
         tf.close()
         os.unlink(tf.name)
